% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balken.mehrfach.plot.R
\name{balken.mehrfach.plot}
\alias{balken.mehrfach.plot}
\title{Balkendiagram von mehrfachen Variabeln untereinander}
\usage{
balken.mehrfach.plot(df, varnames = names(df),
  skala = c("Trifft überhaupt nicht zu", "Trifft überwiegend zu",
  "Trifft eher zu", "Trifft gerade noch zu", "Trifft überwiegend zu",
  "Trifft voll und ganz zu"), out = c("png", "pdf"), filename = "name",
  colors = brewer.pal(5, "Greens"), bg = colors[1])
}
\arguments{
\item{df}{The data-frame that holds all variables}

\item{varnames}{A list of Variable names you like to get printed}

\item{out}{Saving the graph as a file? Then opting in for out="png" or out="pdf"}

\item{filename}{The filename for the graph}

\item{colors}{The color palette. Using RColorBrewer as example}

\item{bg}{The backgound color}
}
\description{
Balkendiagram von mehrfachen Variabeln untereinander
}
\examples{
vars <- c("Vis_Gut_1",
           "Vis_Zurecht_2",
           "Vis_Umg_3",
           "Vis_Spaz_4",
           "Vis_Übers_5",
           "Vis_Stimm_6",
           "Vis_Einsehb_7",
           "Vis_Abw_8",
           "Vis_Schutz_9",
           "Vis_Nat_10")
df <- data.frame(SLS[vars])
#names(SLS)
Vars 1 -10  
filename="Visuelle Bewertung der Wohnumgebung.png"
varnames <- c("Die Umgebeung gefällt mir hier sehr gut.",
             "Ich finde mich hier in der Umgebung gut zurecht.",
             "Ich finde die Umgebung hier ist gut zu überschauen.",
             "Die Umgebung hier lädt zum Spazierengehen ein.",
             "Ich empfinde die Umgebung als sehr übersichtlich.",
             "Ich empfinde die Umgebung als stimmig und organisiert.",
             "Es gibt hier viele schlecht einsehbare Ecken.",
             "Ich empfinde die Umgebung als sehr vielfältig und abwechslungsreich.",
             "Es gibt in der umgebung viele Möglichkeiten, Schutz vor Gefahren zu finden.",
             "Ich denke, die Umgebung hier ist sehr natürlich.")
colors <- brewer.pal(5, "Greens") # colors from RColorBrewer package. 5 Colors
balken.mehrfach.plot(df, varnames=varnames, out = "png", filename=filename)
skala <- c("Trifft überhaupt nicht zu", "Trifft überwiegend zu", "Trifft eher zu", "Trifft gerade noch zu", "Trifft überwiegend zu", "Trifft voll und ganz zu")
}

