% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpDF.R
\name{bpDF}
\alias{bpDF}
\title{bpDF Boxploting one or two whole dataframes with variables and means}
\usage{
bpDF(df, df2 = NULL, df.name = "", df2.name = "",
  bpVarNames = names(df), bpTitle = names(df), bpSubtitle = "Subtitle",
  bpSkala = NULL, nlevels = max(df[1]), colors = brewer.pal(5, "Greens"),
  bg = colors[1], bpCite = paste("© @muuksi", lubridate::today()), ...)
}
\arguments{
\item{df}{a data.frame with variables}

\item{df2}{the second dataframe if desired}

\item{df.name}{A string to plot left for identifiying the data}

\item{df2.name}{For the second dataframe}

\item{bpVarNames}{Variable names. Ideally the whole question}

\item{bpTitle}{The title}

\item{bpSubtitle}{The subtitle}

\item{bpSkala}{The scale: A list with strings}

\item{nlevels}{How many levels have the variables?}

\item{colors}{Colors for means. A list with five colors}

\item{bg}{Color for the background. Defaults to first color in colors}

\item{bpCite}{The source of the Data or copyrigth notice}

\item{...}{Params passing througth to boxplot(...)}
}
\description{
bpDF Boxploting one or two whole dataframes with variables and means
}
\examples{


formulierungen <- c("Ich befürworte es, wenn die Stadt versucht, \\nbei der Straßen- und Gehwegbeleuchtung \\nStrom zu sparen.",
                     "Es ist gut, wenn die Stadt versucht,\\n in den öffentlichen Gebäuden \\nEnergie zu sparen.",
                     "Die Stadt sollte versuchen, \\n sich auf neunen Wegen \\nfür den Klimaschutz einzusetzen.")
bpTitle = "Möglichkeiten der Energieeinsparung"
bpSubtitle = ""
bpSkala = c("Trifft überhaupt nicht zu", "Trifft überwiegend zu", "Trifft eher zu", "Trifft gerade noch zu", "Trifft überwiegend zu", "Trifft voll und ganz zu")
bpCite=paste("© @muuksi",lubridate::today(),"| Quelle: M.Hinse & F.Lüdicke, Saarlouis-Steinrausch // IZES.de")
nlevels=6
df = LED.df

# Funktion aufrufen
bpDF(df, bpVarNames, bpTitle, bpSubtitle="LED", bpSkala, nlevels, color, bpCite)


bpDF(KONV.df, df2 = LED.df, df.name = "HQL", df2.name="LED", bpTitle=bpTitle, bpSubtitle = bpSubtitle, bpVarNames = formulierungen, bpSkala = bpSkala, nlevels = 6, bpCite = bpCite)
}

