% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.plot.bici.R
\name{likert.plot.bici}
\alias{likert.plot.bici}
\title{A custom likert plot
saves a ggplot2 png file in current working directory}
\usage{
likert.plot.bici(df, varnames, nlevels = NA, title = "likert bar plot",
  group = NULL, file, bpCite = paste("© @muuksi", lubridate::today()))
}
\arguments{
\item{df}{dataframe with variables to plot(ideally already factored)}

\item{varnames}{a list with variable names to use}

\item{nlevels}{How many levels do the variables have?}

\item{title}{a title string}

\item{group}{grouping variable}

\item{file}{filename}

\item{bpCite}{not used...}
}
\description{
A custom likert plot
saves a ggplot2 png file in current working directory
}
\examples{


}

