% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndCite.R
\name{loadAndCite}
\alias{loadAndCite}
\title{loadAndCite packages}
\usage{
loadAndCite(packages = c("devtools", "likert", "reshape", "psych",
  "questionr", "foreign", "GPArotation", "knitr", "corrplot"),
  bibtex = "packages.bibtex")
}
\arguments{
\item{packages}{A list of packages to install and cite}

\item{bib}{bibtex file with packages to export}
}
\description{
Load and if not installed yet install packges in the list
}
\examples{
packages <- c("devtools", "likert", "reshape", "psych", "questionr", "foreign", "GPArotation", "knitr", "corrplot")
bib <- "loaded_packages.bibtex"
loadAndCite(packages, bib)

}

