% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.bar.plot.R
\name{percent.bar.plot}
\alias{percent.bar.plot}
\title{percent.bar.plot
       Funktion zur Erstellung einer Prozent BalkenGrafik}
\usage{
\method{percent}{bar.plot}(xx, title, ylab, xlab, pCite = paste("© @muuksi",
  lubridate::today()), statistic = c("bin", "count"), save = TRUE,
  bin = NA)
}
\arguments{
\item{xx}{data as data.frame}

\item{title}{The title}

\item{ylab}{Description Y-axes}

\item{xlab}{Description X-axses}

\item{statistic}{"count" or "bin" only one}

\item{save}{if TRUE a 20x20cm png will be saved}

\item{bin}{binwidth parameter only for "bin"}
}
\description{
percent.bar.plot
       Funktion zur Erstellung einer Prozent BalkenGrafik
}
\examples{
pCite <- paste("© @muuksi",lubridate::today(),"| Quelle: M.Hinse // hinse.eu")
percent.bar.plot(xx=berlin.bikes$frame,  title = "Rahmen ", ylab = "Häufigkeit", 
               xlab = "Material", pCite=paste("© @muuksi",lubridate::today(),"| Quelle: M.Hinse // hinse.eu"), 
               statistic = "count", save=TRUE)
ggsave("Verwendete Rahmen Materialien.png", units="cm", width = 20, height = 15, dpi=300)



}

