% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionsAsVarnames.R
\name{questionsAsVarnames}
\alias{questionsAsVarnames}
\title{Adding the questions form survey to the end of variable names in dataframe}
\usage{
questionsAsVarnames(dataframe, questions)
}
\arguments{
\item{dataframe}{A dataframe containing the data}

\item{questions}{A list with the questions for the new variable names}
}
\value{
new_df
}
\description{
Adding the questions form survey to the end of variable names in dataframe
}
\examples{
# Ausformulierte Fragen angeben. Gleiche Anzahl, gleiche Reihenfolge
head(cars)
questions <- c("How fast is the car?", "How far can you you go with it? ")
new.df <- questionsAsVarnames(dataframe=cars, questions)
head(new.df)

}

